+++ wp-content/themes/ink/content.php
@@ -27,6 +27,15 @@ $categories_list = get_the_category_list( ', ' );
 				</div>
 			<?php endif; ?>
 
+			<?php
+			$sticky_text = stag_theme_mod( 'post_settings', 'sticky_text' );
+
+			if ( is_sticky() && '' !== $sticky_text ) : ?>
+				<div class="sticky-tag">
+					<p><?php echo esc_html( $sticky_text ); ?></p>
+				</div>
+			<?php endif; ?>
+
 			<h1 class="entry-title">
 				<a href="<?php the_permalink(); ?>" rel="bookmark">
 				<?php the_title(); ?>